PDFAnalyzer/requirements.txt
pdfplumber
streamlit
pytesseract
Pillow
google-generativeai
python-dotenv
